Move file to generator.py
